+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased]
 
+### Added
+
+- Recenter (Ctrl-L) can be invoked during Search/replace.
 - New command "Append Next Kill".
 - New commands "Kill Ring Copy" and "Kill Ring Cut".
 - New command "Kill Whole Line".